+++ changelog.md
@@ -6,3 +6,4 @@
 ## 7 September 2023
 * Addition of the `Reset expression and context` button to the test page
 * Removal of the Terminology Server URL from the resource page (still uses the value in the settings page)
+* Add new settings URL for where to retrieve example resources from (changes default from sqlonfhir-r4 to the public HAPI demo server)